Move smooth curves code out of selection module
